Introduce an AggressiveSpeculation pass
This pass can aggressively speculate then/else blocks such that control flow is eliminated but data flow is preserved via "select" instructions.

E.g.,:

    int v;
    if (__builtin_unpredictable(c)) {
      int r = a + 3;
      v = call(r);
    } else {
      v = call(7) + 10;
    }
    return v;

becomes:

    int v;

    int r = a + 3;          // speculative from "then"
    v = call(c ? r : 7);    // merged from both
    int v_false = v + 10;   // speculative from "else"

    return c ? v : v_false;

Because this results in the speculation of entire basic blocks, we're unlikely to be able to get profitability right based solely on static analysis. As a result, this transformation is considered to be profitable only when we have "[unpredictable](https://llvm.org/docs/LangRef.html#unpredictable-metadata)" metadata on the IR, which typically appears as a result of a programmer hint (see above) or upcoming PMU-based LLVM profile feedback.

This initial implementation of the transformation can recognize and merge at most 1 non-speculatable function call somewhere in each of the then/else blocks. This is a simplification to support the first application of interest, and could be made more general in the future.